#MF-22 add navigation for Page500
